(client|server)2.cppを(client|server).cppに変更
